Hn/quick resources page (#134)
* Adding "Quick Resources Page," anchor tags to HiDPI headers in screen-setup (WIP)

Created the "Quick Resources Page." Created anchor tags in HTML in screen-setup for HiDPI related headers.

* Fixed anchor tags for HiDPI headers

Removed the a tags on screen-setup page and fixed the anchor tags on quick-resources page.

* Added Quick Resources to Instructor page; Fixed typo

* Added Quick Resources to Instructor page; Fixed typo

Looking for a way to have "Go back to Top" link on bottom of "Quick Resources" page - possibly Gatsby plugin or JavaScript?

* Update content/instructor/quick-resources/index.mdx

Co-Authored-By: Joel Hooks <joel@egghead.io>

* Change vid embeds to links; Update folder name to "recording-cheatsheet"

Co-authored-by: Hiroko <hiro@24villages.com>
Co-authored-by: Hiro Nishimura <xhirokox@gmail.com>
Co-authored-by: Joel Hooks <joel@egghead.io>